// client/src/app/_models/customer.ts
export class Customer {
  Id: number = 0;
  Name: string = '';
  IndustryId: number = 0;
  AddressId: number = 0;
  VatRegistrationNo: string = '';
  CustomerUsage: string = '';
  BillingAddressId: number = 0;
  IsPurchaseCapacity: boolean = false;
  IsSellCapacity: boolean = false;
  UseBothCapacity: boolean = false;
}